pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tabular-ml-pipeline"
version = "0.1.0"
description = "Reusable scikit-learn pipeline for classic tabular datasets"
authors = [
  {name = "Your Name"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24,<2.0",
  "pandas>=2.0,<3.0",
  "scikit-learn>=1.4,<1.5",
  "pyyaml>=6.0",
  "joblib>=1.3",
  "matplotlib>=3.7",
]

[project.optional-dependencies]
test = [
  "pytest>=7.0",
]

[project.scripts]
ml-project = "ml_project.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
